refactor: シフト行の表示を共通化
